+++ dev/run_all_examples.sh
@@ -1,4 +1,6 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+set -eo pipefail
 
 # Root directory
 root_directory=$(pwd)