+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.7.0] - 2024-02-26
+
+### Added 
+
+- Vault service. Post quantum signing algorithms support
+
+### Changed
+
+- Now targets Java 17 and 21.
+- Rewrote `README.md`.
+
 
 ## [3.6.0] - 2024-01-11
 
@@ -274,7 +284,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Domain Intel client
 - Redact client
 
-[unreleased]: https://github.com/pangeacyber/pangea-java/compare/v3.6.0...main
+[unreleased]: https://github.com/pangeacyber/pangea-java/compare/v3.7.0...main
+[3.7.0]: https://github.com/pangeacyber/pangea-java/compare/v3.6.0...v3.7.0
 [3.6.0]: https://github.com/pangeacyber/pangea-java/compare/v3.5.0...v3.6.0
 [3.5.0]: https://github.com/pangeacyber/pangea-java/compare/v3.4.0...v3.5.0
 [3.4.0]: https://github.com/pangeacyber/pangea-java/compare/v3.3.0...v3.4.0